#19 > folder 구조 변경 > findIdPwd폴더 위치를 /findIdPwd에서 /account/findIdPwd로 변경 - com.kh.workPeople.findIdPwd.controller.FindIdPwdController.java 제거 - AccountController에 findIdPwd관련 핸들러 메소드 선언
