+++ src/main/java/com/kh/workPeople/account/controller/AccountController.java
@@ -1,6 +1,7 @@
 package com.kh.workPeople.account.controller;
 
 import org.springframework.stereotype.Controller;
+import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
 
@@ -31,5 +32,38 @@ public class AccountController {
 	@GetMapping("/changePassword")
 	public String changePassword() {
 		return "account/join/change_password";
+	}
+	
+	@RequestMapping(value={"/findIdPwd/", "/findIdPwd/main"})
+	public String findIdPwdMain() {
+		return "account/findIdPwd/main";
+	}
+	
+	@GetMapping("/findIdPwd/personal/findId")
+	public String personalFindId(Model model) {
+		model.addAttribute("member_type", "personal");
+
+		return "account/findIdPwd/findId";
+	}
+	
+	@GetMapping("/findIdPwd/personal/findPwd")
+	public String personalFindPwd(Model model) {
+		model.addAttribute("member_type", "personal");
+
+		return "account/findIdPwd/findPwd";
+	}
+	
+	@GetMapping("/findIdPwd/company/findId")
+	public String companyFindId(Model model) {
+		model.addAttribute("member_type", "company");
+
+		return "account/findIdPwd/findId";
+	}
+	
+	@GetMapping("/findIdPwd/company/findPwd")
+	public String companyFindPwd(Model model) {
+		model.addAttribute("member_type", "company");
+
+		return "account/findIdPwd/findPwd";
 	}
 }
