+++ src/main/resources/templates/account/findIdPwd/findPwd.html
@@ -4,7 +4,7 @@
 <meta charset="UTF-8">
 <title>workPeople</title>
 <link rel="stylesheet" th:href="@{/css/common/common.css}">
-<link rel="stylesheet" th:href="@{/css/findIdPwd/find-id-pwd.css}">
+<link rel="stylesheet" th:href="@{/css/account/findIdPwd/find-id-pwd.css}">
 </head>
 <body>
 	<div id="container">
@@ -34,6 +34,6 @@
 		</div>
 	</div>
 	<script th:src="@{/js/common/joinValidate.js}"></script>
-	<script th:src="@{/js/findIdpwd/find-id-pwd.js}" ></script>
+	<script th:src="@{/js/account/findIdpwd/find-id-pwd.js}" ></script>
 </body>
 </html>
