+++ src/main/resources/templates/account/findIdPwd/main.html
@@ -4,7 +4,7 @@
 <meta charset="UTF-8">
 <title>workPeople</title>
 <link rel="stylesheet" th:href="@{/css/common/common.css}">
-<link rel="stylesheet" th:href="@{/css/findIdPwd/main.css}">
+<link rel="stylesheet" th:href="@{/css/account/findIdPwd/main.css}">
 </head>
 <body>
 	<div id="container">
@@ -15,14 +15,14 @@
 				<ul>
 					<li>개인회원
 						<ul>
-							<li><a th:href="@{/findIdPwd/personal/findId}">아이디 찾기</a>
-							<li><a th:href="@{/findIdPwd/personal/findPwd}">비밀번호 찾기</a>
+							<li><a th:href="@{/account/findIdPwd/personal/findId}">아이디 찾기</a>
+							<li><a th:href="@{/account/findIdPwd/personal/findPwd}">비밀번호 찾기</a>
 						</ul>
 					</li>
 					<li>기업회원
 						<ul>
-							<li><a th:href="@{/findIdPwd/company/findId}">아이디 찾기</a>
-							<li><a th:href="@{/findIdPwd/company/findPwd}">비밀번호 찾기</a>
+							<li><a th:href="@{/account/findIdPwd/company/findId}">아이디 찾기</a>
+							<li><a th:href="@{/account/findIdPwd/company/findPwd}">비밀번호 찾기</a>
 						</ul>
 					</li>
 				</ul>
