+++ src/main/resources/templates/account/login/login.html
@@ -19,7 +19,7 @@
 			<div th:replace="account/login/common-login.html"></div>
 			<div class="signup-forgotten">
 				<p><a th:href="@{/account/join/personal-join}">회원가입</a></p>
-				<p><a th:href="@{/findIdPwd/main}">아이디/비밀번호 찾기</a></p>
+				<p><a th:href="@{/account/findIdPwd/main}">아이디/비밀번호 찾기</a></p>
 				<p><a th:href="@{/account/manager-login}">관리자 로그인</a></p>
 			</div>
 			<div class="sns_login">
